requirements.txt
# ============================================================
# OCR 系统精简依赖列表（优化打包体积）
# 版本：v1.3.1
# 说明：已移除 PaddleOCR/RapidOCR Python 版本，使用 C++ 引擎
# ============================================================

# GUI 框架
PySide6>=6.6.0

# 图像处理（核心）
Pillow>=10.0.0
# pillow-heif>=0.14.0  # HEIF/HEIC 格式支持（未使用，已移除）

# Excel 导出
openpyxl>=3.1.0

# PDF 处理
PyMuPDF>=1.23.0

# 基础数值计算（可选，仅在线OCR引擎处理numpy数组时需要）
# numpy>=1.24.0,<2.0.0  # 已移除：系统不直接使用numpy，仅在线OCR引擎支持numpy数组输入

# 阿里云 OCR（新版 SDK）
alibabacloud-ocr-api20210707>=3.1.2
alibabacloud-tea-openapi>=0.3.0
alibabacloud-tea-util>=0.3.0

# DeepSeek OCR（硅基流动平台）
openai>=1.0.0

# ============================================================
# 已移除的大型依赖（使用 C++ 引擎替代）：
# - paddleocr (~300MB) → 使用 PaddleOCR-json.exe
# - paddlepaddle (~500MB+) → 使用 PaddleOCR-json.exe  
# - rapidocr-onnxruntime (~100MB) → 使用 RapidOCR-json.exe
# - opencv-python (~80MB) → 项目未使用
# - aliyun-python-sdk-* → 已升级到新版 SDK
# 
# 预计减少打包体积：~1GB
# ============================================================
